+++ docs/conf.py
@@ -40,7 +40,7 @@ project = "COMPAS"
 copyright = "Block Research Group - ETH Zurich"
 author = "Tom Van Mele"
 
-release = "1.15.0"
+release = "1.15.1"
 version = ".".join(release.split(".")[0:2])
 
 master_doc = "index"