+++ src/compas/geometry/primitives/polyline.py
@@ -413,3 +413,103 @@ class Polyline(Primitive):
         return division_pts
 
     divide_by_length = divide_polyline_by_length
+
+    def extend(self, length):
+        """Extends a polyline by a given length, by modifying the first and/or last point tangentially.
+
+        Parameters:
+        -----------
+        length: float or tuple[float, float]
+            A single length value to extend the polyline only at the end,
+            or two length values to extend at both ends.
+
+        Returns
+        -------
+        None
+
+        """
+        try:
+            start, end = length
+            self.points[0] = self.points[0] + self.lines[0].vector.unitized().scaled(-start)
+        except TypeError:
+            start = end = length
+        self.points[-1] = self.points[-1] + self.lines[-1].vector.unitized().scaled(end)
+
+    def extended(self, length):
+        """Returns a copy of this polyline extended by a given length.
+
+        Parameters:
+        -----------
+        length: float or tuple[float, float]
+            A single length value to extend the polyline only at the end,
+            or two length values to extend at both ends.
+
+        Returns
+        -------
+        :class:`~compas.geometry.Polyline`
+
+        """
+        crv = self.copy()
+        crv.extend(length)
+        return crv
+
+    def shorten(self, length):
+        """Shortens a polyline by a given length.
+
+        Parameters:
+        -----------
+        length: float or tuple[float, float]
+            A single length value to shorten the polyline only at the end,
+            or two length values to shorten at both ends.
+
+        Returns
+        -------
+        None
+
+        """
+        try:
+            start, end = length
+            total_length = 0
+            for line in self.lines:
+                total_length += line.length
+                if total_length < start:
+                    del self.points[0]
+                elif total_length == start:
+                    del self.points[0]
+                    break
+                else:
+                    self.points[0] = line.end + line.vector.unitized().scaled(-(total_length-start))
+                    break
+        except TypeError:
+            start = end = length
+
+        total_length = 0
+        for i in range(len(self.lines)):
+            line = self.lines[-(i+1)]
+            total_length += line.length
+            if total_length < end:
+                del self.points[-1]
+            elif total_length == end:
+                del self.points[-1]
+                break
+            else:
+                self.points[-1] = line.start + line.vector.unitized().scaled(total_length-end)
+                break
+
+    def shortened(self, length):
+        """Returns a copy of this polyline shortened by a given length.
+
+        Parameters:
+        -----------
+        length: float or tuple[float, float]
+            A single length value to shorten the polyline only at the end,
+            or two length values to shorten at both ends.
+
+        Returns
+        -------
+        :class:`~compas.geometry.Polyline`
+
+        """
+        crv = self.copy()
+        crv.shorten(length)
+        return crv