+++ src/compas_plotters/__init__.py
@@ -14,7 +14,7 @@ compas_plotters
 
 """
 
-__version__ = '1.15.0'
+__version__ = '1.15.1'
 
 from .core import *  # noqa: F401 F403
 # from .artists import *  # noqa: F401 F403