+++ tests/compas/datastructures/test_mesh.py
@@ -7,6 +7,7 @@ from compas.datastructures import Mesh
 from compas.datastructures import meshes_join_and_weld
 from compas.geometry import Box
 from compas.geometry import Polygon
+from compas.geometry import Polyhedron
 from compas.geometry import Translation
 from compas.geometry import allclose
 
@@ -182,6 +183,25 @@ def test_to_vertices_and_faces():
     assert len(vertices) == 36
     assert len(faces) == 25
 
+
+def test_to_vertices_and_faces_triangulated():
+    # tri
+    mesh = Mesh.from_shape(Polyhedron.from_platonicsolid(4))
+    vertices, faces = mesh.to_vertices_and_faces(triangulated=True)
+    assert len(vertices) == 4
+    assert len(faces) == 4
+
+    # quad
+    mesh = Mesh.from_shape(Polyhedron.from_platonicsolid(6))
+    vertices, faces = mesh.to_vertices_and_faces(triangulated=True)
+    assert len(vertices) == 8
+    assert len(faces) == 12
+
+    # ngon
+    mesh = Mesh.from_shape(Polyhedron.from_platonicsolid(12))
+    vertices, faces = mesh.to_vertices_and_faces(triangulated=True)
+    assert len(vertices) == 32
+    assert len(faces) == 60
 
 # --------------------------------------------------------------------------
 # helpers