+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Added `Polyline.extend`, `Polyline.extended`, `Polyline.shorten`,  `Polyline.shortened`.
-* Added `Data.hash` for computing a hash value of data objects, for example for comparisons during version control.
+* Added `Data.sha256` for computing a hash value of data objects, for example for comparisons during version control.
 
 ### Changed