+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-* Fixed bug in `normal_polygon` in `compas.geometry._core`.
+* Fixed bug in `normal_polygon` in `compas.geometry`.
 
 ### Removed